feat(register): move test functionality to `buba/register`
Current buba funcionality is moved to buba/register to make way for
a source code compilation feature.
