+++ test/mocha.opts
@@ -1,3 +1,3 @@
- --require ./index.js
+ --require ./register/index.js
  --reporter spec
  --ui tdd
